+++ application/config/routes.php
@@ -54,3 +54,5 @@ $route['404_override'] = '';
 $route['translate_uri_dashes'] = FALSE;
 $route['profile'] = 'portfolio';
 $route['profile/(:any)'] = 'portfolio/one/$1';
+$route['stock'] = 'stock';
+$route['stock/(:any)'] = 'stock/history/$1';