+++ application/controllers/Portfolio.php
@@ -1,13 +1,7 @@
 <?php
 /**
- * NEED TO ASK --- How to do current holdings since Donald sells
- * one of his holdings but don't know initial value...lame
- * Was thinking of adding and substracting to get total
- * value for each holding and if no transactions, they have 0 in that
- * holding.
- * Requirement is "Current holdings should show the quantity
- * held for each stock on the BSX"
- *  -Kamil (Just something to think about)
+ * This is the portfolio(profile) controller used to pass data
+ * to display profile information.
  */
 
 class Portfolio extends Application {
@@ -22,7 +16,7 @@ class Portfolio extends Application {
 
         parent::__construct();
 
-        $this->load->model('portfolios');
+        $this->load->model('transactions');
         $this->load->model('players');
     }
 
@@ -32,7 +26,7 @@ class Portfolio extends Application {
 
     /**
      * Shows the player whose currently logged in and if not logged,
-     * shows ____(something), right now just shows all player transactions
+     * shows Donald. Uses a different pagebody for a user profile as
      */
     function index() {
 
@@ -40,16 +34,16 @@ class Portfolio extends Application {
         if(!isset($_SESSION['user'])) {
             $this->data['pageTitle'] = 'Profile of Donald' ;
             // this is the view we want shown
-            $this->data['pagebody'] = 'user_profile';
+            $this->data['pagebody'] = 'profile';
 
             //Data to fill in dropdown menu
             $this->data['player_names'] = $this->players->all();
 
             //Data to fill transactions table, Donald for now
-            $this->data['players'] = $this->portfolios->some("Player", 'Donald');
+            $this->data['players'] = $this->transactions->some("Player", 'Donald');
 
             //Player's current holdings in each stock, Donald for now
-            $this->data['stocks']= $this->portfolios->get_player_stocks("Player", 'Donald');
+            $this->data['stocks']= $this->transactions->get_player_stocks("Player", 'Donald');
 
             $this->render();
         }
@@ -69,13 +63,13 @@ class Portfolio extends Application {
         $this->data['pagebody'] = 'profile';
 
         //Player's current holdings in each stock
-        $this->data['stocks']= $this->portfolios->get_player_stocks("Player", $player);
+        $this->data['stocks']= $this->transactions->get_player_stocks("Player", $player);
 
         //Data to fill in dropdown menu
         $this->data['player_names'] = $this->players->all();
 
         //Data to fill transactions table
-        $this->data['players'] = $this->portfolios->some("Player", $player);
+        $this->data['players'] = $this->transactions->some("Player", $player);
 
         $this->render();
     }